check for number of args
git-svn-id: svn+ssh://svn.wikia-code.com/svn/wikia/trunk@21714 e846bd91-c042-42a7-a218-4b65e2c1904d

+++ extensions/wikia/WebsiteWiki/thumbnail.sh
@@ -6,14 +6,17 @@
 #
 # usage thumbnail.sh <url> <target.png>
 
+if [ $# -ne 2 ]
+then
+  echo "Usage: `basename $0` <url> <target.png>"
+  exit $E_BADARGS
+fi
 TMPFILE="`tempfile`".png
 URL="$1"
 TARGET="$2"
-#echo $TMPFILE;
-echo "Getting $URL"
- /usr/bin/xvfb-run --server-args="-screen 0, 1024x768x24" /usr/bin/cutycapt --url=$URL --out=$TMPFILE --delay=5000
+/usr/bin/xvfb-run --server-args="-screen 0, 1024x768x24" /usr/bin/cutycapt --url=$URL --out=$TMPFILE --delay=5000
 if [ -f $TMPFILE ];
 then
-	/usr/bin/convert -thumbnail "250x188^" -crop 250x188+0+0 -antialias +repage $TMPFILE $TARGET
+	/usr/bin/convert -thumbnail "250x188^" -crop "250x188+0+0" +repage $TMPFILE $TARGET
 	rm $TMPFILE
 fi
